+++ doc/release_notes.rst
@@ -4,7 +4,9 @@ Release notes
 * Module ximpol.core.fitsio added (issue #49).
 * ximpol.evt.event refactored to use the new ximpol.core.fitsio module.
 * GTI list in the output event file (issue #24)
-* ROI table in the output event file (issue #45).
+* ROI source table in the output event file (issue #45).
+* IRF name added in the output event file header (issue #24).
+* ROI information added in the output event file header (issue #48).
 
 
 *ximpol (0.15.2) - Mon, 25 Jan 2016 18:04:33 -0800*